pyproject.toml
[tool.poetry]
name = "hat-py-sdk"
version = "0.25.11"
description = ""
authors = ["rtatton <rdt17@proton.me>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/rtatton/hat-py-sdk"
repository = "https://github.com/rtatton/hat-py-sdk"
keywords = ["dataswift", "hat", "python", "sdk"]
packages = [{ include = "hat", from = "src" }]

[tool.poetry.dependencies]
python = "^3.7"

PyJWT = { extras = ["crypto"], version = "^2.4.0" }
aiodns = { optional = true, version = "^3.0.0" }
aiohttp = "^3.8.1"
aiohttp-client-cache = { optional = true, version = "^0.7.3" }
cchardet = { optional = true, version = "^2.1.7" }
humps = "^0.2.2"
keyring = "^23.8.2"
orjson = { optional = true, version = "^3.7.12" }
pydantic = "^1.9.1"
python-ulid = { optional = true, version = "^1.1.0" }
requests = "^2.28.1"
requests-cache = "^0.9.5"

[tool.poetry.extras]
all = ["aiodns", "aiohttp-client-cache", "cchardet", "orjson", "python-ulid"]
fast = ["aiodns", "aiohttp-client-cache", "cchardet", "orjson"]
ulid = ["python-ulid"]

[tool.poetry.dev-dependencies]
black = "^22.6.0"
coverage = { extras = ["toml"], version = "^6.4.4" }
darglint = "^1.8.1"
flake8 = "^5.0.2"
flake8-black = "^0.3.3"
flake8-bugbear = "^22.7.1"
isort = "^5.10.1"
mypy = "^0.971"
nox = "^2022.8.7"
nox-poetry = "^1.0.1"
pep8-naming = "^0.13.1"
pre-commit = "^2.20.0"
pre-commit-hooks = "^4.3.0"
pytest = "^7.1.2"
pyupgrade = "^2.37.3"
safety = "^2.1.1"
typeguard = "^2.13.3"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["hat", "tests"]

[tool.coverage.report]
fail_under = 0
show_missing = true

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict = true
warn_unreachable = true


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
